Issue #9598 & #8697: Fixes force expand crash
